+++ tests/test_visitors/test_ast/test_blocks/conftest.py
@@ -9,9 +9,7 @@ multiple_assign = '{0} = unmatched_assign = 1'
 annotated_assign1 = '{0}: type = 1'
 annotated_assign2 = '{0}: type'
 unpacking_assign1 = '{0}, unmatched_assign = (1, 2)'
-unpacking_assign2 = 'unmatched_assign1, {0}, unmatched_assign2 = (1, 2)'
-unpacking_assign3 = '{0}, *unmatched_assign2 = (1, 2)'
-unpacking_assign4 = 'unmatched_assign, *{0} = (1, 2)'
+unpacking_assign2 = 'unmatched_assign, *{0} = (1, 2)'
 
 
 @pytest.fixture(params=[
@@ -21,8 +19,6 @@ unpacking_assign4 = 'unmatched_assign, *{0} = (1, 2)'
     annotated_assign2,
     unpacking_assign1,
     unpacking_assign2,
-    unpacking_assign3,
-    unpacking_assign4,
 ])
 def assign_statement(request):
     """Parametrized fixture that contains all possible assign templates."""