+++ tests/test_visitors/test_ast/test_blocks/test_try_block.py
@@ -189,8 +189,7 @@ def test_except_block_usage(
     try_template9,
 ])
 @pytest.mark.parametrize('first_name, second_name', [
-    ('unique_name', 'unique_name1'),
-    ('unique_name', '_unique_name'),
+    ('unique_name', 'unique_name2'),
 ])
 def test_except_block_correct(
     assert_errors,