+++ tests/test_visitors/test_ast/test_naming/test_naming_rules/test_wrong_names.py
@@ -9,7 +9,11 @@ from wemake_python_styleguide.violations.naming import (
 from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
 
 
-@pytest.mark.parametrize('wrong_name', VARIABLE_NAMES_BLACKLIST)
+@pytest.mark.parametrize('wrong_name', [
+    'value',
+    'no',
+    'data',
+])
 def test_wrong_variable_name(
     assert_errors,
     assert_error_text,