+++ tests/test_visitors/test_ast/test_naming/test_variable_usages.py
@@ -1,24 +1,29 @@
 # -*- coding: utf-8 -*-
 
-import string
-
 import pytest
 
-from wemake_python_styleguide.visitors.ast.naming import (
-    VARIABLE_NAMES_BLACKLIST,
-    WrongNameVisitor,
-)
+from wemake_python_styleguide.visitors.ast.naming import WrongNameVisitor
+
+annotation = 'some_var: {0}'
+annotation_value = 'some_var: {0} = None'
+assigned = 'some_var = {0}'
+assigned_attribute = '{0}.attribute = 1'
 
 import_name = 'import {0}'
 from_import_module = 'from {0} import some'
 from_import_name = 'from some import {0}'
 
 calling_function = 'print({0})'
+calling_star_function = 'print(*{0})'
 called_function = '{0}()'
 calling_method = 'instance.call({0})'
 called_method = 'instance.{0}()'
 accessing_attribute = 'instance.{0}'
 accessed_attribute = '{0}.attribute'
+
+key_access = 'instance[{0}]'
+list_definition = 'instance = [{0}, 1]'
+
 raising_variable = 'raise {0}'
 
 returning_variable = """
@@ -36,10 +41,12 @@ def function():
     yield {0}
 """
 
+inheriting_variables = 'class ValidName({0}): ...'
+
 
 @pytest.mark.parametrize('bad_name', [
-    *VARIABLE_NAMES_BLACKLIST,
-    *string.ascii_letters,
+    'value',  # blacklisted
+    'x',  # short
     '__Class_private',
     'number_prefix_10',
     'some__underscores',
@@ -49,19 +56,27 @@ def function():
     'wrong_alias_',
 ])
 @pytest.mark.parametrize('code', [
+    annotation,
+    annotation_value,
+    assigned,
+    assigned_attribute,
     import_name,
     from_import_module,
     from_import_name,
     calling_function,
+    calling_star_function,
     called_function,
     calling_method,
     called_method,
     accessing_attribute,
     accessed_attribute,
+    key_access,
+    list_definition,
     raising_variable,
     returning_variable,
     awaiting_variable,
     yielding_variable,
+    inheriting_variables,
 ])
 def test_wrong_variable_names(
     assert_errors,