+++ wemake_python_styleguide/logic/naming/name_nodes.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import ast
-from typing import Optional
+from typing import List, Optional
 
 
 def is_same_variable(left: ast.AST, right: ast.AST) -> bool:
@@ -26,4 +26,32 @@ def get_assigned_name(node: ast.AST) -> Optional[str]:
     if isinstance(node, ast.ExceptHandler):
         return getattr(node, 'name', None)
 
+    return None
+
+
+def get_variables_from_node(node: ast.AST) -> List[str]:
+    """Gets the assigned names from the list of nodes."""
+    names: List[str] = []
+    naive_attempt = _extract_name(node)
+
+    if naive_attempt:
+        names.append(naive_attempt)
+    elif isinstance(node, ast.Tuple):
+        for subnode in node.elts:
+            extracted_name = _extract_name(subnode)
+            if extracted_name:
+                names.append(extracted_name)
+    return names
+
+
+def _extract_name(node: ast.AST) -> Optional[str]:
+    """
+    Utility to extract names for several types of nodes.
+
+    Should not be used direclty, use safer :py:`~get_assign_names` function.
+    """
+    if isinstance(node, ast.Starred):
+        node = node.value
+    if isinstance(node, ast.Name):
+        return node.id
     return None